separate component + fix update akun
