+++ src/screens/InfoAkun/styles.js
@@ -2,20 +2,6 @@ import {StyleSheet} from 'react-native';
 import {COLORS, FONTS, SIZES} from '../../themes';
 
 export default StyleSheet.create({
-  icon: {
-    opacity: 0.7,
-    alignItems: 'center',
-    alignSelf: 'center',
-    justifyContent: 'center',
-  },
-  container: {
-    height: 100,
-    width: 100,
-    borderRadius: 15,
-    alignSelf: 'center',
-    backgroundColor: '#E2D4F0',
-    justifyContent: 'center',
-  },
   form: {
     marginHorizontal: SIZES.padding,
     marginTop: 24,
