+++ src/store/actions/index.js
@@ -1,2 +1,5 @@
-export * from './user';
+export * from './akun';
 export * from './auth';
+export * from './common';
+export * from './home';
+export * from './notification';
